Support multiple listeners per-process
Use select(2) to multiplex non-blocking accept(2) calls between
them.  Additionally, aggressively make a bet after accepting
clients where we'll try to do a non-blocking accept(2) against
the full set of descriptors.  This is based on the assumption
that if we just accepted connections, we're probably reasonably
busy.

This should lead to lower latency under high load; but some
wasted cycles when requests come in intermitently.  By this same
logic, we don't really care for the thundering herd problem,
either; since it is only noticeable with many (hundreds) of
processes when most of them are idle.