Use read(2) and blocking I/O for clients
Unicorn is only designed for fast internal networks (and
loopback); so avoid wasting time with userspace I/O buffering.

This should not significantly affect userspace threading on 1.8
in case your application itself is running threads for some
(masochistic) reason as long as the clients you're serving
directly with Unicorn are fast.