add add_listener method for use in configs
This makes it possible to bind per-process listener ports for
easier debugging.  One of my biggest gripes about other prefork
webservers is that strace-ing the correct process for debugging
is difficult.  This makes it possible for each worker to bind to
a unique port or UNIX socket independent of the other workers.