rename http11 => unicorn/http11
Avoid conflicting with existing (and future) Mongrel installs in
case either changes.  Of course, this also allows us more
freedom to experiment and break the API if needed...
However, I'm only planning on making minor changes to
remove the amount of C code we have to maintain and
possibly some minor performance improvements.