Socket: add {snd,rcv}buf opts to bind_listen
bind_listen takes a hash as its second parameter now, allowing
the addition of :sndbuf and :rcvbuf options to specify the size
of the buffers in bytes.  These correspond to the SO_SNDBUF and
SO_RCVBUF options via setsockopt(2) respectively.

This also adds support for per-listener backlogs to be used.

However, this is only an internal API change and the changes
have not yet been exposed to the user via Unicorn::Configurator,
yet.

Also add a bunch of SocketHelper tests