Deferred log rotation in workers
Instead of rotating logs immediately when SIGUSR1 is caught,
defer it until the current client is processing is complete.

This allows multi-line log messages generated by apps to not be
broken up if SIGUSR1 is received while the app is running.

If we're sleeping inside IO.select, we close a pipe in the
exceptfds set to cause EBADF to be raised.

This also adds a small reliability improvement to test_exec
so we wait until signals are ready before sending USR1
to rotate logs.