ensure responses always have the "Status:" header
There are weird (and possibly broken) clients out there that
require it despite being present in the first line of the
response.  So be nice and accomodate them.  Keep in mind that
the Rack SPEC explicitly forbids this header from being in the
headers returned by the Rack-based application; so we have to
always inject it ourselves and ignore it if the application
sets it.