test_upload: ensure StringIO objects are binary
Sockets always return binary encoded data, so when
StringIO.new(string) is called, that StringIO object inherits
the encoding of the initial string it was created with.

And yes, Ruby 1.9 still makes me seriously uncomfortable with
I/O manipulation since the encoding layer does things behind my
back.  UNIX is (and should always be) just a bag of bytes!

Signed-off-by: Eric Wong <normalperson@yhbt.net>