http_response: minor performance gains
Avoid creating garbage every time we lookup the status code
along with the message.  Also, we can use global const arrays
for a little extra performance because we only write one-at-a
time

Looking at MRI 1.8, Array#join with an empty string argument is
slightly better because it skips an append for every iteration.