No point in unsetting the O_NONBLOCK flag
Since we've switched to readpartial, we'll already be protected
from any unpleasant errors that might get thrown at us.  There's
no easy way to prevent MRI from calling a select() internally to
check for readiness, so speculative+blocking read() calls are
out already.

Additionally, most requests come in the form of GETs which are
fully-buffered in the kernel before we even accept() the socket;
so a single readpartial call will be enough to fully consume it.