Instant shutdown signals really mean instant shutdown
Use SIGQUIT if you're going to be nice and do graceful
shutdowns.  Sometimes people run real applications on this
server and SIGINT/SIGTERM get lost/trapped when Object is
rescued and that is not good.  Also make sure we break out of
the loop properly when the master is dead.

Testcases added for both SIGINT and dead master handling.