Unicorn::App::Inetd: reinventing Unix, poorly :)
This includes an example of tunneling the git protocol inside a
TE:chunked HTTP request.  The example is unfortunately contrived
in that it relies on the custom examples/cat-chunk-proxy.rb
script in the client.  My initial wish was to have a generic
tool like curl(1) operate like this:

  cat > ~/bin/cat-chunk-proxy.sh <<EOF
  #!/bin/sh
  exec curl -sfNT- http://$1:$2/
  EOF
  chmod +x ~/bin/cat-chunk-proxy.sh
  GIT_PROXY_COMMAND=cat-chunk-proxy.sh git clone git://0:8080/foo

Unfortunately, curl will attempt a blocking read on stdin before
reading the TCP socket; causing the git-clone consumer to
starve.  This does not appear to be a problem with the new
server code for handling chunked requests.