http: make strings independent before modification
Ruby strings may be copy-on-write (multiple #dup'ed strings can
point to the same internal buffers).  So since
HttpParser#headers modifies its buffer argument, we'll need to
make sure the String object we have points to its own private
buffer.

This doesn't affect Unicorn except in a to-be-written test case
that relies on a #dup'ed String.