http: cleanup and avoid potential signedness warning
Just pass the http_parser struct pointer when checking for
invalid headers in the trailer.  The compiler should be smart
enough to inline and not relookup the flags.  This avoids having
to worry about the flags being signed or not (they should never
be) and also makes it easier to maintain if we move away from
using bitfields.