README: emphasize the "fast clients"-only part
While Unicorn is one of very many Unix-only, pre-forking, shared
socket servers in existence, and Unicorn is _definitely_ not the
only server that only works *well* with fast clients, either.

But as far as we know, Unicorn is the first (and so far only)
server that emphasizes only working well with fast clients.