cleanup: avoid redundant error checks for fstat
If fstat() fails on an open file descriptor in the master,
something is seriously wrong (like your kernel is broken/buggy)
and trying to restart the worker that owned that file descriptor
is likely masking the symptoms.  Instead let the error propagate
up to the main loop to avoid wasting cycles to restart broken
workers.