preserve user/group ownership when reopening logs
This is only supported when SIGUSR1 is sent only to the master
process (which then resends SIGUSR1 to the workers).

Since we only added support for user/group switching in the
workers, we now chown any log files upon switching users so the
master can pick up and chown the log files later on.  Thus
we can avoid having to restart workers because they fail to
rotate log files on their own.