test_server: ensure stderr is written to before reading
This works around a race condition caused by the server
closing the connection before writing out to stderr in
the ensure block.  So to ensure we've waited on the server
to write to the log file, just send another HTTP request
since we know our test server only processes requests
serially.