http: allow userinfo component in absoluteURIs
This is not explicitly specified or listed as an example in in
rfc2616.  However, rfc2616 section 3.2.1 defers to rfc2396[1]
for the definition of absolute URIs, so the userinfo component
should be allowable, even if it does not make any sense.

In the real world, previous versions of Mongrel used URI.parse()
and thus allowed userinfo, so we also have precedence to allow
userinfo to be compatible *in case* our interpretation of the
RFCs is incorrect.  This change is unfortunately needed because
*occasionally* real clients rely on them.

Reported-by: Scott Chacon

[1] rfc3986 obsoletes rfc2396, but also includes userinfo