http: pedantic fix for trailer-less chunked requests
HTTP requests without trailers still need a CRLF after the last
chunk, that is: it must end as: "0\r\n\r\n", not "0\r\n".  So
we'll always pretend there are trailers to parse for the
sake of TeeInput.

This is mostly a pedantic fix, as the two bytes in the socket
buffer are unlikely to trigger protocol errors.