update test infrastructure to run Rubinius tests
In parallel with other of Rubies, of course.  We need to rely on
RUBY_ENGINE since RUBY_VERSION is 1.8.7 and that conflicts with
the most popular MRI version.

Since Rubinius doesn't support some command-line options, we
still need to rely on MRI for a few things.  Also fixing an
embarrassing UUoC in the process.