cleanup makefiles, restore aggregation screwup
The publish_doc target belongs in main, since other people
may update the website.

local.mk.sample enables a subset of tests for Rubinius
and allows them to run in parallel with the MRI tests.

And it was NOT a UUoC after all, there are multiple files to
aggregate :x