workaround rbx not reopening logs to stderr/stdout
While log reopening worked reliably for newly-created File
objects in the unit tests, the $stderr and $stdout handles that
get redirected did not get reopened reliably under Rubinius.

We work around this by relying on Rubinius internals and
directly setting the @path instance variable.  This is harmless
for MRI and should be harmless for other any other Ruby
implementations we'll eventually support.

ref: http://github.com/evanphx/rubinius/issues/360